Rename LinkOverlayYAML to TaskDefinitionViewer and move to separate file
